+++ packages/tailwind/src/components/drawer/snippet.tsx
@@ -1,3 +1,4 @@
+import { ark } from '@ark-ui/react'
 import * as Ark from '@ark-ui/react/dialog'
 import { createStyleContext } from '~/lib/create-style-context'
 import { drawerStyles } from './recipe'
@@ -14,6 +15,9 @@ export const DrawerContent = withContext(Ark.Dialog.Content, 'content')
 export const DrawerDescription = withContext(Ark.Dialog.Description, 'description')
 export const DrawerTitle = withContext(Ark.Dialog.Title, 'title')
 export const DrawerTrigger = withContext(Ark.Dialog.Trigger, 'trigger')
+export const DrawerHeader = withContext(ark.div, 'header')
+export const DrawerBody = withContext(ark.div, 'body')
+export const DrawerFooter = withContext(ark.div, 'footer')
 
 export const Drawer = Object.assign(DrawerRoot, {
   Root: DrawerRoot,
@@ -24,4 +28,7 @@ export const Drawer = Object.assign(DrawerRoot, {
   Description: DrawerDescription,
   Title: DrawerTitle,
   Trigger: DrawerTrigger,
+  Header: DrawerHeader,
+  Body: DrawerBody,
+  Footer: DrawerFooter,
 })