+++ packages/tailwind/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+- Added Header, Footer and Body to Drawer
+- Added sm and lg variants to Slider
+
 ## [0.12.2] - 2023-10-24
 
 ### Added